add version guard for cleanup with nodes < 6.6

+++ test/framework/src/main/java/org/elasticsearch/test/rest/ESRestTestCase.java
@@ -410,7 +410,12 @@ public abstract class ESRestTestCase extends ESTestCase {
             waitForPendingRollupTasks();
         }
 
-        if (hasXPack && false == preserveILMPoliciesUponCompletion()) {
+        /*
+         * ILM was introduced in 6.6 so any cluster that contains older
+         * nodes won't be able to do *anything* with ILM, including cleanup.
+         */
+        if (hasXPack && nodeVersions.first().onOrAfter(Version.V_6_6_0)
+                && false == preserveILMPoliciesUponCompletion()) {
             deleteAllPolicies();
         }
     }
